move auto conversion test into tests
