+++ crates/nu-command/tests/commands/database/into_sqlite.rs
@@ -495,3 +495,16 @@ fn insert_test_rows(dirs: &Dirs, nu_table: &str, sql_query: Option<&str>, expect
 
     assert_eq!(expected, actual_rows);
 }
+
+#[test]
+fn test_auto_conversion() {
+    Playground::setup("sqlite json auto conversion", |_, playground| {
+        let raw = "{a_record:{foo:bar,baz:quux},a_list:[1,2,3],a_table:[[a,b];[0,1],[2,3]]}";
+        nu!(cwd: playground.cwd(), "{} | into sqlite filename.db -t my_table", raw);
+        let outcome = nu!(
+            cwd: playground.cwd(),
+            "open filename.db | get my_table.0 | to nuon --raw"
+        );
+        assert_eq!(outcome.out, raw);
+    });
+}
